Alien Invision with Hand Tracking
I have written this game in Python using Pygame, CV2 and Mediapipe modules. The game is like this, the webcam is turned on and the background of the game is the image displayed by the webcam. In the game, you move a ship by moving your hand, which shoots by opening and closing your hand, and your goal is to destroy the spaceships before they reach you. Difficulty level of the game is in such a way that in each level, the speed of the enemies increases and after level 3, the webcam image is reversed.